Pickups
* Bullet Pickup Prefab and Materials Created

Pickup prefabs for the ammo of the SMG and Pistol weapons were created. These can be found in Prefabs folder labeled as PistolAmmoPickups and SMGAmmoPickups. These prefabs were created using a variety of cylinders and capsules to create a replica of a bullet. A box collider was also created which will be used later. The materials used to colour these prefabs are in the Materials folder under the Bullets folder. One is a metallic gold and the other is a metallic silver. These prefabs have been placed in the scene near the start for further testing and trialing.

* Implemented Ammo Pickup Spawner

A PickupSpawner C# script was created which spawns a select pickup over time which can be changed within Unity. An AmmoPickups C# script was also created which allow the player to trigger with the Pickup box colliders and start the spawn countdown for the spawner. A basic debug log lets the player know which pickup they have collected and how much ammo it will increase by. Once the weapons are implemented this can be added to the weapon script to actually be functional. The grenades will also be implemented later and a feature to update the UI will be implemented later.

* Implemented Pickup Rotation and Floating Animation

A Rotate C# script was created that rotates the pickup around the y-axis by a changeable speed and hovers the pickup up and down smoothly by implementing a sine graph that the pickup transforms along. The height and speed of the hovering are all changeable within Unity. The AmmoPickups script was renames to Pickups so that the grenade can be implemented within this script once a prefab has been developed.

* Grenade Prefab Developed and Implemented

A Grenade prefab was created to act as a pickup item to add to the player's grenade inventory. This was implemented following the same format for the ammo pickups. A new grenade material of a dark metallic grey was created and the Rotate and Pickups C# scripts were added to the prefab.  A grenade spawner was also created in the scene similar to the ammo spawners.

* Pickups Refined and Health Pickups Added

A health pickup prefab was created with a corresponding spawner in the scene. The pickup types as an enum were given their own C# script and the Pickups Scripts uses the name space PickupScripts which is the folder in where all of the C# scripts relating to the pickups.