FEAT: Player UI
* text mesh pro added

* added uimanager

* Player UI Basic Implementation

A canvas was re-created with a slider for player health, and some text for the ammo and magazine amounts. The UIManager C# script was attached to the Player UI canvas and the corresponding Serialize Fields within the UIManager were added to the component from within the canvas.

* Player UI Prefab Created

The PlayerUI in the scene was turned into a prefab in the prefabs folder. The PlayerUI prefab was then added directly to the player and everything works perfectly. The prefab could also be added to the scene and this would give the same effect.

---------

Co-authored-by: MelodyMad <M_M337@student.usc.edu.au>