FEAT: Health for player & enemies
* Player Health Bar Implemented

The health bar for the player was implemented through the PlayerUI prefab attached to the Player. A HealthManager Script was created to act as an interface for the player and enemy health. A PlayerHealth script was also created which was attached to the player to update the player health bar when damage is taken or a healing pickup is collided with. The player health bar has just been implemented and needs further work still to be able to update when taken damage and when healing. However so far everything seems to be working.

* Enemy Health Bar Implemented

An enemy health bar was created by creating a prefab called EnemyUI in the Enemy prefabs folder, which displays the enemy's health bar. An EnemyHealth script was also created very similar to the PlayerHealth script however it is for the enemies. EnemyHealth inherits from the HealthManager as well. Everything seems to be running smoothly so far, however further work is still required to make sure that the enemy can take damage from the player's bullets.

* Damage to Enemy Implemented

The bullet script was slightly altered by calling the TakeDamage method from the HealthManager when the bullet collides with the Enemy when it has the EnemyHealth script attached. The EnemyHealth script was split into a new script called EnemyHealthBar which focuses solely on the health bar, whereas the EnemyHealth focuses on damage, healing, and death. The EnemyHealthBar script allows the health bar above the enemy to rotate along with the players camera so that it is always easily in view in the scene. The EnemyHealthBar script also updates the health bar and in the EnemyHealth script it references the health bar as a child of the enemy and calls the UpdateHealth method.

---------

Co-authored-by: MelodyMad <M_M337@student.usc.edu.au>