FEAT: start & pause menus
* Start Menu Implemented

A Start Menu was created so that the player can play the game or exit on the main menu. A new scene was created called StartMenu and a C# script called StartGameMenu was created so that when the play button was clicked it would start playing the game on the MainLevel and if the quit button was clicked then the game would exit. In the StartMenu these buttons are present on a canvas and there are some objects for decoration within the scene.

* Pause Menu Implemented

A new prefab called PauseMenu was created which controls the canvas and the buttons to resume the game or quit the game. A C# script called PauseMenuScreen was created which controls the code for when the buttons are clicked and when ESC is pressed to bring up the pause menu. The PlayerMovement script was slightly altered by adding and if statement within the code for the camera. Therefore, when the menu is paused, the camera will not move with the mouse when interacting with the menu, as it returns all of the line of code for the camera. When the game is no longer paused the camera will function as normal.

* Update StartMenu Buttons

There was a small error with the buttons. The starting menu works when you are in the StartMenu scene, which when play is clicked you play in the MainLevel. For the paused menu to work, you must implement the PauseMenu prefab into the desired scene.

---------

Co-authored-by: MelodyMad <M_M337@student.usc.edu.au>