Commenting and Ideas
A few scripts were commented slightly to further explain certain areas to improve readability and understanding. In the WeaponsManager an idea for the implementation of the health pickup was commented and an idea to update the player UI was also commented. In the UIManager the Reload and RoundShot scripts were commented for future evaluation and clean up as these methods are similar to what is in the WeaponsManager.