GameOver Menu
* Game Over Death Screen Implemented

A GameOverScreen prefab was created which includes a red tinted background, some text, and two buttons to replay the game from the MainLevel or to exit the game by returning to the StartMenu. A C# script was created called GameOver and is similar to the PausedMenuScreen script. The GameOverScreen will be called when the player runs out of health and dies. In the PlayerMovement script, in the camera movement with the mouse the game over screen was added as an exception to stop the camera movement when the game is over. I also created a PlayerInterface prefab that has the Player, UI, PausedMenu, and GameOverScreen attached. So that all the menus are in one place and you only have to add one prefab to the scene rather than two.

* added Pause action bound to ESC key

* Creating Static Methods for Other Scripts to Access

The GameOver and PausedMenuScreen C# scripts were improved to include static methods. Rather than using update to check if the screen should be activated a static method was created that is called from other scripts to perform the method. GameOver has a static TriggerGameOver method which is called in the Player health that when the player dies to perform that method. In the PausedMenuScreen the TogglePause method was transformed into a static method that is called from the PlayerInput when ESC is pressed on the keyboard. In the StartGameMenu the StartGame() and GoToStartGame() became static so that the GameOver and PausedMenuScreen can call those methods directly. Everything works perfectly and runs smoothly.

* needed EventsSystem

---------

Co-authored-by: MelodyMad <M_M337@student.usc.edu.au>